clean up and organize Jpeg test utilities
