+++ tests/ImageSharp.Tests/Formats/Jpg/Block8x8FTests.cs
@@ -1,24 +1,23 @@
 ﻿// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-using System.Diagnostics;
 
-using SixLabors.ImageSharp.Formats;
-using SixLabors.ImageSharp.Formats.Jpeg.Common;
-using SixLabors.ImageSharp.Formats.Jpeg.GolangPort.Components;
-using SixLabors.ImageSharp.Formats.Jpeg.GolangPort.Utils;
-
-using Xunit;
-using Xunit.Abstractions;
 
 // Uncomment this to turn unit tests into benchmarks:
 //#define BENCHMARKING
 
 // ReSharper disable InconsistentNaming
 
-namespace SixLabors.ImageSharp.Tests
+namespace SixLabors.ImageSharp.Tests.Formats.Jpg
 {
     using System;
+    using System.Diagnostics;
+
+    using SixLabors.ImageSharp.Formats.Jpeg.Common;
+    using SixLabors.ImageSharp.Tests.Formats.Jpg.Utils;
+
+    using Xunit;
+    using Xunit.Abstractions;
 
     public class Block8x8FTests : JpegUtilityTestFixture
     {
@@ -222,88 +221,7 @@ namespace SixLabors.ImageSharp.Tests
             sw.Stop();
             this.Output.WriteLine($"TranposeInto_PinningImpl_Benchmark finished in {sw.ElapsedMilliseconds} ms");
         }
-
-        [Fact]
-        public void iDCT2D8x4_LeftPart()
-        {
-            float[] sourceArray = Create8x8FloatData();
-            float[] expectedDestArray = new float[64];
-
-            ReferenceImplementations.iDCT2D8x4_32f(sourceArray, expectedDestArray);
-
-            Block8x8F source = new Block8x8F();
-            source.LoadFrom(sourceArray);
-
-            Block8x8F dest = new Block8x8F();
-
-            DCT.IDCT8x4_LeftPart(ref source, ref dest);
-
-            float[] actualDestArray = new float[64];
-            dest.CopyTo(actualDestArray);
-
-            this.Print8x8Data(expectedDestArray);
-            this.Output.WriteLine("**************");
-            this.Print8x8Data(actualDestArray);
-
-            Assert.Equal(expectedDestArray, actualDestArray);
-        }
-
-        [Fact]
-        public void iDCT2D8x4_RightPart()
-        {
-            float[] sourceArray = Create8x8FloatData();
-            float[] expectedDestArray = new float[64];
-
-            ReferenceImplementations.iDCT2D8x4_32f(sourceArray.AsSpan().Slice(4), expectedDestArray.AsSpan().Slice(4));
-
-            Block8x8F source = new Block8x8F();
-            source.LoadFrom(sourceArray);
-
-            Block8x8F dest = new Block8x8F();
-
-            DCT.IDCT8x4_RightPart(ref source, ref dest);
-
-            float[] actualDestArray = new float[64];
-            dest.CopyTo(actualDestArray);
-
-            this.Print8x8Data(expectedDestArray);
-            this.Output.WriteLine("**************");
-            this.Print8x8Data(actualDestArray);
-
-            Assert.Equal(expectedDestArray, actualDestArray);
-        }
-
-        [Theory]
-        [InlineData(1)]
-        [InlineData(2)]
-        [InlineData(3)]
-        public void TransformIDCT(int seed)
-        {
-            Span<float> sourceArray = Create8x8RandomFloatData(-200, 200, seed);
-            float[] expectedDestArray = new float[64];
-            float[] tempArray = new float[64];
-
-            ReferenceImplementations.iDCT2D_llm(sourceArray, expectedDestArray, tempArray);
-
-            // ReferenceImplementations.iDCT8x8_llm_sse(sourceArray, expectedDestArray, tempArray);
-            Block8x8F source = new Block8x8F();
-            source.LoadFrom(sourceArray);
-
-            Block8x8F dest = new Block8x8F();
-            Block8x8F tempBuffer = new Block8x8F();
-
-            DCT.TransformIDCT(ref source, ref dest, ref tempBuffer);
-
-            float[] actualDestArray = new float[64];
-            dest.CopyTo(actualDestArray);
-
-            this.Print8x8Data(expectedDestArray);
-            this.Output.WriteLine("**************");
-            this.Print8x8Data(actualDestArray);
-            Assert.Equal(expectedDestArray, actualDestArray, new ApproximateFloatComparer(1f));
-            Assert.Equal(expectedDestArray, actualDestArray, new ApproximateFloatComparer(1f));
-        }
-
+        
         [Fact]
         public unsafe void CopyColorsTo()
         {
@@ -365,74 +283,6 @@ namespace SixLabors.ImageSharp.Tests
             {
                 Assert.InRange(val, 0, 255);
             }
-        }
-
-        [Theory]
-        [InlineData(1)]
-        [InlineData(2)]
-        public void FDCT8x4_LeftPart(int seed)
-        {
-            Span<float> src = Create8x8RandomFloatData(-200, 200, seed);
-            Block8x8F srcBlock = new Block8x8F();
-            srcBlock.LoadFrom(src);
-
-            Block8x8F destBlock = new Block8x8F();
-
-            float[] expectedDest = new float[64];
-
-            ReferenceImplementations.fDCT2D8x4_32f(src, expectedDest);
-            DCT.FDCT8x4_LeftPart(ref srcBlock, ref destBlock);
-
-            float[] actualDest = new float[64];
-            destBlock.CopyTo(actualDest);
-
-            Assert.Equal(actualDest, expectedDest, new ApproximateFloatComparer(1f));
-        }
-
-        [Theory]
-        [InlineData(1)]
-        [InlineData(2)]
-        public void FDCT8x4_RightPart(int seed)
-        {
-            Span<float> src = Create8x8RandomFloatData(-200, 200, seed);
-            Block8x8F srcBlock = new Block8x8F();
-            srcBlock.LoadFrom(src);
-
-            Block8x8F destBlock = new Block8x8F();
-
-            float[] expectedDest = new float[64];
-
-            ReferenceImplementations.fDCT2D8x4_32f(src.Slice(4), expectedDest.AsSpan().Slice(4));
-            DCT.FDCT8x4_RightPart(ref srcBlock, ref destBlock);
-
-            float[] actualDest = new float[64];
-            destBlock.CopyTo(actualDest);
-
-            Assert.Equal(actualDest, expectedDest, new ApproximateFloatComparer(1f));
-        }
-
-        [Theory]
-        [InlineData(1)]
-        [InlineData(2)]
-        public void TransformFDCT(int seed)
-        {
-            Span<float> src = Create8x8RandomFloatData(-200, 200, seed);
-            Block8x8F srcBlock = new Block8x8F();
-            srcBlock.LoadFrom(src);
-
-            Block8x8F destBlock = new Block8x8F();
-
-            float[] expectedDest = new float[64];
-            float[] temp1 = new float[64];
-            Block8x8F temp2 = new Block8x8F();
-
-            ReferenceImplementations.fDCT2D_llm(src, expectedDest, temp1, downscaleBy8: true);
-            DCT.TransformFDCT(ref srcBlock, ref destBlock, ref temp2, false);
-
-            float[] actualDest = new float[64];
-            destBlock.CopyTo(actualDest);
-
-            Assert.Equal(actualDest, expectedDest, new ApproximateFloatComparer(1f));
         }
 
         [Theory]
