+++ tests/ImageSharp.Tests/Formats/Jpg/JpegProfilingBenchmarks.cs
@@ -1,18 +1,18 @@
 // Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-using System;
-using System.IO;
-using System.Linq;
-using System.Numerics;
-using SixLabors.ImageSharp.Formats;
-using SixLabors.ImageSharp.Formats.Jpeg;
-using SixLabors.ImageSharp.PixelFormats;
-using Xunit;
-using Xunit.Abstractions;
-
-namespace SixLabors.ImageSharp.Tests
+namespace SixLabors.ImageSharp.Tests.Formats.Jpg
 {
+    using System;
+    using System.IO;
+    using System.Linq;
+    using System.Numerics;
+
+    using SixLabors.ImageSharp.Formats.Jpeg;
+
+    using Xunit;
+    using Xunit.Abstractions;
+
     public class JpegProfilingBenchmarks : MeasureFixture
     {
         public JpegProfilingBenchmarks(ITestOutputHelper output)
