+++ tests/ImageSharp.Tests/Formats/Jpg/JpegUtilsTests.cs
@@ -1,16 +1,18 @@
 // Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-using System.Numerics;
 
-using SixLabors.ImageSharp.Formats.Jpeg.GolangPort.Utils;
-using SixLabors.ImageSharp.PixelFormats;
-
-using Xunit;
 
 // ReSharper disable InconsistentNaming
-namespace SixLabors.ImageSharp.Tests
+namespace SixLabors.ImageSharp.Tests.Formats.Jpg
 {
+    using System.Numerics;
+
+    using SixLabors.ImageSharp.Formats.Jpeg.GolangPort.Utils;
+    using SixLabors.ImageSharp.PixelFormats;
+
+    using Xunit;
+
     public class JpegUtilsTests
     {
         public static Image<TPixel> CreateTestImage<TPixel>()
