+++ tests/ImageSharp.Tests/Formats/Jpg/LibJpegToolsTests.cs
@@ -1,8 +1,9 @@
-namespace SixLabors.ImageSharp.Tests
+namespace SixLabors.ImageSharp.Tests.Formats.Jpg
 {
     using System.IO;
 
     using SixLabors.ImageSharp.PixelFormats;
+    using SixLabors.ImageSharp.Tests.Formats.Jpg.Utils;
 
     using Xunit;
