+++ tests/ImageSharp.Tests/ImageSharp.Tests.csproj
@@ -40,5 +40,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="TestUtilities\Factories\" />
+    <Folder Include="Utils\Jpg\Formats\" />
   </ItemGroup>
 </Project>
